프로젝트 관리 CLI 종료 방식 변경
`sys.exit(0)`을 사용하여 프로젝트 관리 CLI를 종료합니다.
이는 더 안정적인 종료 방식을 제공합니다.
